historian/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
strawberry-graphql==0.268.1
influxdb-client==1.48.0
jinja2==3.1.6
